pyproject.toml
[project]
name = "rag-pdf-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.124.4",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.4",
    "numpy>=2.3.5",
    "openai>=2.13.0",
    "openai-agents>=0.6.3",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
